+++ packages/terra-core-docs/src/terra-dev-site/doc/tag/Tag.1.doc.mdx
@@ -1,8 +1,8 @@
 import { Badge } from 'terra-tag/package.json?dev-site-package';
 
-import TagDefault from '../example/TagDefault?dev-site-example';
-import TagHref from '../example/HrefTag?dev-site-example';
-import TagFallbacks from '../example/TagFallbacks?dev-site-example';
+import TagDefault from './example/TagDefault?dev-site-example';
+import TagHref from './example/HrefTag?dev-site-example';
+import TagFallbacks from './example/TagFallbacks?dev-site-example';
 
 import TagPropsTable from 'terra-tag/src/Tag?dev-site-props-table';
