+++ packages/terra-core-docs/src/terra-dev-site/test/tag/DefaultTag.test.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import classNames from 'classnames/bind';
+import Tag from 'terra-tag';
+import styles from './TagTestCommon.module.scss';
+
+const cx = classNames.bind(styles);
+const DefaultTag = () => (
+  <div className={cx('tag-wrapper')}>
+    <Tag
+      id="default"
+      text="Default Tag"
+    />
+  </div>
+);
+export default DefaultTag;
