+++ packages/terra-core-docs/src/terra-dev-site/test/tag/HrefTag.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Tag from '../../../Tag';
+import Tag from 'terra-tag';
 
 const HrefTag = () => <Tag id="href" text="Href Tag" href="www.google.com" />;
