+++ packages/terra-tag/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Migrate docs to terra-core-docs.
+
 ## 2.51.1 - (January 4, 2022)
 
 * Changed
